# packages/cactus-plugin-dpp/src/main/json/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hyperledger Cacti Plugin - Digital Product Passport (DPP)",
    "description": "API for Digital Product Passport interoperability - Fundão Cherry Case Study. Implements full lifecycle from production to consumer feedback, including disaggregation, cross-chain transfers via SATP, and role-based supply-chain operations.",
    "version": "2.0.0",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/api/v1/@hyperledger/cactus-plugin-dpp/create": {
      "post": {
        "summary": "Farmer: Create initial DPP NFT",
        "operationId": "createDPP",
        "tags": ["Farmer"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDPPRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DPP Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateDPPResponse" }
              }
            }
          },
          "500": {
            "description": "Creation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/@hyperledger/cactus-plugin-dpp/data": {
      "get": {
        "summary": "Get DPP on-chain data by ID",
        "operationId": "getDPPData",
        "tags": ["Public"],
        "parameters": [
          {
            "name": "dppId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "DPP data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetDPPDataResponse" }
              }
            }
          },
          "500": {
            "description": "Fetch failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/@hyperledger/cactus-plugin-dpp/history/{dppId}": {
      "get": {
        "summary": "Audit: Get full provenance history of a DPP",
        "operationId": "getDPPHistory",
        "tags": ["Audit"],
        "parameters": [
          {
            "name": "dppId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical trail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetDPPHistoryResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/@hyperledger/cactus-plugin-dpp/transfer": {
      "post": {
        "summary": "Transfer ownership to another actor",
        "operationId": "transferDPP",
        "tags": ["Ownership"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransferDPPRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ownership transferred",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/@hyperledger/cactus-plugin-dpp/amend": {
      "post": {
        "summary": "Amend DPP metadata URI",
        "operationId": "amendDPPData",
        "tags": ["Processor"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AmendDPPDataRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadata amended",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/@hyperledger/cactus-plugin-dpp/add-certification": {
      "post": {
        "summary": "Attach a certification to a DPP",
        "operationId": "addCertification",
        "tags": ["Processor"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddCertificationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Certification added",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/@hyperledger/cactus-plugin-dpp/update-transport-data": {
      "post": {
        "summary": "Transporter: Record transport leg data",
        "operationId": "updateTransportData",
        "tags": ["Logistics"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTransportDataRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transport data logged",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/@hyperledger/cactus-plugin-dpp/update-retail-data": {
      "post": {
        "summary": "Retailer: Update retail-specific data",
        "operationId": "updateRetailData",
        "tags": ["Retailer"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateRetailDataRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retail data updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/@hyperledger/cactus-plugin-dpp/submit-product-review": {
      "post": {
        "summary": "Consumer: Submit product review/feedback",
        "operationId": "submitProductReview",
        "tags": ["Consumer"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmitProductReviewRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review submitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenericResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/@hyperledger/cactus-plugin-dpp/disaggregate": {
      "post": {
        "summary": "Split a DPP into multiple independent child DPPs",
        "operationId": "disaggregateDPP",
        "tags": ["Processor"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DisaggregateDPPRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Disaggregation successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DisaggregateDPPResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Error message" }
        }
      },
      "GenericResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "txHash": { "type": "string" }
        }
      },
...
      "CreateDPPRequest": {
        "type": "object",
        "required": ["owner"],
        "properties": {
          "owner": { "type": "string", "description": "Ethereum address of the DPP owner" },
          "productionData": { "$ref": "#/components/schemas/ProductionData" }
        }
      },
      "ProductionData": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" },
          "origin": { "type": "string" },
          "productionMethod": { "type": "string" },
          "variety": { "type": "string" },
          "calibre": { "type": "string" },
          "brixDegree": { "type": "string" },
          "certifications": { "type": "array", "items": { "type": "string" } },
          "manufacturer": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "ipfsUri": { "type": "string" }
        }
      },
      "CreateDPPResponse": {
        "type": "object",
        "properties": {
          "dppId": { "type": "string" },
          "tokenId": { "type": "string" },
          "txHash": { "type": "string" },
          "owner": { "type": "string" }
        }
      },
      "GetDPPDataResponse": {
        "type": "object",
        "properties": {
          "dppData": { "$ref": "#/components/schemas/DPPData" }
        }
      },
      "DPPData": {
        "type": "object",
        "properties": {
          "dppId": { "type": "string" },
          "productId": { "type": "string" },
          "productName": { "type": "string" },
          "status": { "type": "string", "enum": ["ACTIVE", "IN_TRANSIT", "RECEIVED", "RETAIL_READY", "BURNED", "LOCKED_CROSSCHAIN", "REVOKED", "UNKNOWN"] },
          "creationDate": { "type": "string" },
          "owner": { "type": "string" },
          "publicData": { "type": "object", "description": "Parsed JSON metadata" },
          "certifications": { "type": "array", "items": {} }
        }
      },
...
      "GetDPPHistoryResponse": {
        "type": "object",
        "properties": {
          "history": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HistoryRecord" }
          }
        }
      },
      "HistoryRecord": {
        "type": "object",
        "properties": {
          "event": { "type": "string" },
          "actor": { "type": "string" },
          "timestamp": { "type": "integer", "description": "Unix timestamp" },
          "locationFrom": { "type": "string" },
          "locationTo": { "type": "string" },
          "conditionData": { "type": "string" },
          "certification": { "type": "object" },
          "childDppId": { "type": "string" },
          "inheritedFrom": { "type": "string" }
        }
      },
      "TransferDPPRequest": {
        "type": "object",
        "required": ["dppId", "from", "to"],
        "properties": {
          "dppId": { "type": "string" },
          "from": { "type": "string", "description": "Current owner address" },
          "to": { "type": "string", "description": "New owner address" }
        }
      },
      "AmendDPPDataRequest": {
        "type": "object",
        "required": ["dppId"],
        "properties": {
          "dppId": { "type": "string" },
          "handlerAddress": { "type": "string" },
          "newData": { "type": "object", "description": "Key-value pairs to amend in the DPP metadata" }
        }
      },
      "AddCertificationRequest": {
        "type": "object",
        "required": ["dppId", "certificationData"],
        "properties": {
          "dppId": { "type": "string" },
          "handlerAddress": { "type": "string" },
          "certificationData": { "type": "object", "description": "Certification object (certId, issuedBy, date, etc.)" }
        }
      },
      "UpdateTransportDataRequest": {
        "type": "object",
        "required": ["dppId"],
        "properties": {
          "dppId": { "type": "string" },
          "transportData": {
            "type": "object",
            "description": "Transport leg details",
            "properties": {
              "locationFrom": { "type": "string" },
              "locationTo": { "type": "string" },
              "timestamp": { "type": "string" },
              "conditionData": { "type": "string" },
              "handler": { "type": "string" },
              "handlerAddress": { "type": "string" }
            }
          }
        }
      },
...
      "UpdateRetailDataRequest": {
        "type": "object",
        "required": ["dppId"],
        "properties": {
          "dppId": { "type": "string" },
          "handlerAddress": { "type": "string" },
          "handler": { "type": "string" },
          "location": { "type": "string" },
          "arrivalDate": { "type": "string" },
          "shelfLife": { "type": "string" },
          "price": { "type": "string" },
          "retailData": {
            "type": "object",
            "description": "Alternative nested retail data object",
            "properties": {
              "location": { "type": "string" },
              "arrivalDate": { "type": "string" },
              "shelfLife": { "type": "string" },
              "price": { "type": "string" }
            }
          }
        }
      },
      "SubmitProductReviewRequest": {
        "type": "object",
        "required": ["dppId"],
        "properties": {
          "dppId": { "type": "string" },
          "handlerAddress": { "type": "string" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" }
        }
      },
...
      "DisaggregateDPPRequest": {
        "type": "object",
        "required": ["dppId", "count", "handlerAddress"],
        "properties": {
          "dppId": { "type": "string", "description": "Token ID of the DPP to split" },
          "count": { "type": "integer", "minimum": 2, "maximum": 20, "description": "Number of new DPPs to create" },
          "handler": { "type": "string", "description": "Human-readable handler name" },
          "handlerAddress": { "type": "string", "description": "Ethereum address of the handler" }
        }
      },
      "DisaggregateDPPResponse": {
        "type": "object",
        "properties": {
          "newTokenIds": { "type": "array", "items": { "type": "string" }, "description": "Token IDs of the newly created DPPs" },
          "txHash": { "type": "string" }
        }
      },
...
    }
  }
}